feat: fix hook autoload file path
