+++ package-info.xml
@@ -37,15 +37,25 @@
 		<code>installCheck.php</code>
 		<database>install.php</database>
 		<require-dir name="Themes" destination="$boarddir">Theme files</require-dir>
-		<require-dir name="Sources" destination="$sourcedir">Source files</require-dir>
-		<hook hook="integrate_admin_areas" function="$sourcedir/AutoRespondAdmin.php|\AutoRespond\AutoRespondAdmin::menu#"  />
-		<hook hook="integrate_post2_end" function="$sourcedir/AutoRespond.php|\AutoRespond\AutoRespond::createRespond#"  />
+		<require-dir name="Sources/AutoRespond" destination="$sourcedir">Source files</require-dir>
+		<hook
+				hook="integrate_admin_areas"
+				function="$sourcedir/AutoRespond/AutoRespondAdmin.php|\AutoRespond\AutoRespondAdmin::menu#"  />
+		<hook
+			hook="integrate_admin_areas"
+			function="$sourcedir/AutoRespond/AutoRespondAdmin.php|\AutoRespond\AutoRespondAdmin::menu#"  />
+		<hook
+			hook="integrate_post2_end"
+			function="$sourcedir/AutoRespond/AutoRespond.php|\AutoRespond\AutoRespond::createRespond#"  />
 		<redirect url="$scripturl?action=admin;area=autorespond" timeout="4000">redirect.txt</redirect>
 	</install>
-
 	<uninstall for="2.1 - 2.1.99">
-		<hook hook="integrate_admin_areas" function="$sourcedir/AutoRespondAdmin.php|\AutoRespond\AutoRespondAdmin::menu#" reverse="true" />
-		<hook hook="integrate_post2_end" function="$sourcedir/AutoRespond.php|\AutoRespond\AutoRespond::createRespond#" reverse="true" />
+		<hook
+			hook="integrate_admin_areas"
+			function="$sourcedir/AutoRespond/AutoRespondAdmin.php|\AutoRespond\AutoRespondAdmin::menu#" reverse="true" />
+		<hook
+			hook="integrate_post2_end"
+			function="$sourcedir/AutoRespond/AutoRespond.php|\AutoRespond\AutoRespond::createRespond#" reverse="true" />
 		<database>install.php</database>
 		<remove-dir name="$sourcedir/AutoRespond" />
 	</uninstall>
